splitpicker_backend/requirements.txt
firebase_admin==6.3.0
Flask==3.0.0
Flask_Cors==4.0.0
python-dotenv==1.0.0
sshtunnel==0.4.0
